requirements.txt
accelerate==1.6.0
datasets==3.6.0
cryptography==44.0.3
flask==3.1.0
mysql-connector-python==9.3.0
pandas==2.2.3
pymysql==1.1.1
python-dotenv==1.1.0
rapidfuzz==3.13.0
scikit-learn==1.6.1
selenium==4.32.0
spacy==3.8.5
tabulate==0.9.0
transformers==4.51.3
torch==2.7.0+cu128
waitress==3.0.2